refactor code for removing circular import issue
