+++ apps/quickbooks_online/models.py
@@ -524,8 +524,6 @@ class QBOExpense(models.Model):
                 payee_type = 'Debit Card Misc' if destination_attribute.attribute_type == 'BANK_ACCOUNT' else 'Credit Card Misc'
                 # In case credit card account is selected, we need to create credit card misc vendor
                 if payee_type == 'Credit Card Misc':
-                    from apps.quickbooks_online.tasks import get_or_create_misc_vendor
-                    
                     qbo_credentials = QBOCredential.get_active_qbo_credentials(expense_group.workspace_id)
                     qbo_connection = QBOConnector(credentials_object=qbo_credentials, workspace_id=expense_group.workspace_id)
                     entity = get_or_create_misc_vendor(False, qbo_connection)
